Fixed being able to set the vis to vubars mode while WA2.65 mode is active
winamp 2.65 only had the vubars in the windowshade mode, and I guess I forgot the vis menu when I editted how many modes it has when wa2.65 mode is active.

Also optimised the check for wa5/wa265 modes in currentmode so the file could be just a tad bit smaller.